REFACTOR routes template setup
