+++ app.js
@@ -8,8 +8,10 @@ const cookieParser = require("cookie-parser");
 const app = express();
 
 require("./config/database");
-const index = require("./routes/index");
-const beers = require("./routes/beers");
+
+const api = require("./routes/api");
+// const index = require("./routes/index");
+// const beers = require("./routes/beers");
 const handleGlobalError = require("./middlewares/handleGlobalError");
 
 if (process.env.NODE_ENV === "development") {
@@ -24,8 +26,8 @@ app.use(helmet());
 
 //이부분은 나중에 token 판별하는 로직 꽂아넣으면 좋을듯 for auth
 //app.all("*", verifyToken);
-app.use("/", index);
-app.use("/beers", beers);
+app.use("/api", api);
+// app.use("/beers", beers);
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
