+++ routes/api/auth/controller.js
@@ -1,0 +1,5 @@
+const getIndex = (req, res, next) => {
+  res.json("BIBIBIBIBIBNBONONON");
+};
+
+module.exports = { getIndex };
