+++ routes/api/auth/index.js
@@ -1,0 +1,7 @@
+const router = require("express").Router();
+
+const { getIndex } = require("./controller");
+
+router.route("/").get(getIndex);
+
+module.exports = router;
