+++ routes/api/beers/index.js
@@ -1,6 +1,6 @@
 const router = require("express").Router();
 
-const { getBeer } = require("./controllers/beers.controller");
+const { getBeer } = require("./controller");
 
 router.route("/:id").get(getBeer);
