--- routes/controllers/index.controller.js
@@ -1,7 +1,0 @@
-// 폴더 구조 예제를 위한 파일입니다. 삭제 후 진행하시면 됩니다.
-
-const getIndex = (req, res, next) => {
-  res.send("Hello bibino!");
-};
-
-exports.getIndex = getIndex;
